pyproject.toml
[project]
name = "plant-counter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopandas>=1.1.3",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "rasterio>=1.5.0",
    "scikit-learn>=1.8.0",
    "shapely>=2.1.2",
]
